Optimizations: changed the type representing a piece a byte instead of int, optimized array copies in Board using Buffer.BlockCopy

+++ ChessDotNet/Data/ChessPiece.cs
@@ -1,6 +1,6 @@
 ﻿namespace ChessDotNet.Data
 {
-    using Piece = System.Int32;
+    using Piece = System.Byte;
 
     public static class ChessPiece
     {
